# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pending registration saved. User/account/categories are created after email verification.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RegisterResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Start registration and send an email verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/VerifyEmailResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Verify registration code and create the wallet account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailOnlyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Resend the registration verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Authenticate user and return a JWT token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailOnlyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Send a password reset code by email",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Reset password using an email OTP code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout the current user (stateless JWT)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/account/me": {
      "get": {
        "operationId": "AccountController_getMyAccount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AccountDetailsDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the signed-in user's account details",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/account/qr": {
      "get": {
        "operationId": "AccountController_getQrCode",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AccountQrDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a QR code for receiving transfers into this account",
        "tags": [
          "Account"
        ]
      }
    },
...
    "/api/categories": {
      "get": {
        "operationId": "CategoryController_list",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by category type",
            "schema": {
              "type": "string",
              "enum": [
                "income",
                "expense"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of active (non-deleted) categories",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CategoryResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active categories for the current user",
        "tags": [
          "Category"
        ]
      },
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CategoryResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a custom category",
        "tags": [
          "Category"
        ]
      }
    },
    "/api/categories/{id}": {
      "put": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename a category — protected names (Transfer In / Out) are blocked",
        "tags": [
          "Category"
        ]
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a category — protected names (Transfer In / Out) are blocked",
        "tags": [
          "Category"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sonsam Saving API",
    "description": "Personal Finance Tracker — REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Success",
            "description": "Human-readable message"
          },
          "data": {
            "type": "object",
            "description": "Response payload (null when there is no body)"
          }
        },
        "required": [
          "statusCode",
          "message",
          "data"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "message": {
            "type": "string",
            "example": "Verification code sent to your email. Code expires in 10 minutes."
          }
        },
        "required": [
          "email",
          "message"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "securepass123",
            "minLength": 8
          },
          "phone": {
            "type": "string",
            "example": "012345678"
          }
        },
        "required": [
          "fullName",
          "email",
          "password"
        ]
      },
      "UserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          }
        },
        "required": [
          "id",
          "fullName",
          "email"
        ]
      },
      "AccountSummaryDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string",
            "example": "ACC-20260429-00001"
          },
          "balance": {
            "type": "string",
            "example": "0.00"
          }
        },
        "required": [
          "accountNumber",
          "balance"
        ]
      },
      "VerifyEmailResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT access token (expires in 7 days)"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummaryDto"
          },
          "account": {
            "$ref": "#/components/schemas/AccountSummaryDto"
          }
        },
        "required": [
          "token",
          "user",
          "account"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "format": "email"
          },
          "code": {
            "type": "string",
            "example": "482913",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "EmailOnlyDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT access token (expires in 7 days)"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummaryDto"
          }
        },
        "required": [
          "token",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "securepass123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "format": "email"
          },
          "code": {
            "type": "string",
            "example": "482913",
            "minLength": 6,
            "maxLength": 6
          },
          "newPassword": {
            "type": "string",
            "example": "newsecurepass123",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "code",
          "newPassword"
        ]
      },
      "AccountDetailsDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string",
            "example": "ACC-20260429-00001"
          },
          "balance": {
            "type": "string",
            "example": "0.00",
            "description": "Decimal serialized as string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-04-29T04:43:29.568Z"
          }
        },
        "required": [
          "accountNumber",
          "balance",
          "createdAt"
        ]
      },
      "AccountQrDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string",
            "example": "ACC-20260429-00001"
          },
          "ownerName": {
            "type": "string",
            "example": "John Doe"
          },
          "qrImageBase64": {
            "type": "string",
            "example": "data:image/png;base64,iVBORw0KGgoAAAA...",
            "description": "Base64-encoded PNG of the account-number QR code"
          }
        },
        "required": [
          "accountNumber",
          "ownerName",
          "qrImageBase64"
        ]
      },
...
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "13",
            "description": "Category ID (BigInt as string)"
          },
          "name": {
            "type": "string",
            "example": "Freelance"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "income"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-04-29T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "createdAt"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Freelance",
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "income"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Side Hustle",
            "maxLength": 100
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}